fleshed out shadow elem and changed event listener handling
